+++ VoiceOver Designer/Features/Sources/Editor/EditorView.swift
@@ -99,7 +99,6 @@ class EditorView: FlippedView {
         let scrollViewVisibleHeight = scrollView.frame.height - scrollView.contentInsets.verticals
         return scrollViewVisibleHeight / image.size.height
     }
-    
 }
 
 class FlippedView: NSView {
@@ -117,24 +116,5 @@ class FlippedStackView: NSStackView {
 extension NSEdgeInsets {
     var verticals: CGFloat {
         top + bottom
-    }
-}
-
-class CenteredClipView: NSClipView
-{
-    override func constrainBoundsRect(_ proposedBounds: NSRect) -> NSRect {
-        var rect = super.constrainBoundsRect(proposedBounds)
-        if let containerView = self.documentView as? NSView {
-            
-            if (rect.size.width > containerView.frame.size.width) {
-                rect.origin.x = (containerView.frame.width - rect.width) / 2
-            }
-            
-            if(rect.size.height > containerView.frame.size.height) {
-                rect.origin.y = (containerView.frame.height - rect.height) / 2
-            }
-        }
-        
-        return rect
     }
 }
